Move tests over to tests folder again
